Rename Course controller file
